// src/models/contracts.ts
/**
 * Ravimoshark Contract API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: dev@ravimoshark.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Machines } from "./machines";

export interface Contracts {
    refContract?: number;
    machines?: Array<Machines>;
    client?: string;
    identification?: string;
    dateDebut?: string;
    dateFin?: string;
    reconduction?: string;
    loyer?: number;
    miniconso?: number;
    reviewed?: boolean;
    valid?: boolean;
}