// src/models/refContract.ts
/**
 * Ravimoshark Contract API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: dev@ravimoshark.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Contract's reference.
 */
export interface RefContract {
    refContract?: number;
}